Move query methods onto the relation instead of the class.
This will enable us to start chaining relations together like this:

```ruby
Library.find('test:123').books.where(author:
'Salinger').find('test:231')
```
(this doesn't yet work, but this reorganization is prerequisite for that
functionality)
