+++ spec/unit/query_spec.rb
@@ -16,10 +16,12 @@ describe ActiveFedora::Base do
   end
 
   describe ":all" do
+    before { ActiveFedora::Base.stub(:relation => relation) }
     describe "called on a concrete class" do
+      let(:relation) { ActiveFedora::Relation.new(SpecModel::Basic) }
       it "should query solr for all objects with :has_model_s of self.class" do
-        SpecModel::Basic.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
-        SpecModel::Basic.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
+        relation.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
+        relation.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
         mock_docs = [{"id" => "changeme:30"}, {"id" => "changeme:22"}]
         mock_docs.should_receive(:has_next?).and_return(false)
         ActiveFedora::SolrService.instance.conn.should_receive(:paginate).with(1, 1000, 'select', :params=>{:q=>@model_query, :qt => 'standard', :sort => [@sort_query], :fl=> 'id', }).and_return('response'=>{'docs'=>mock_docs})
@@ -27,9 +29,10 @@ describe ActiveFedora::Base do
       end
     end
     describe "called without a specific class" do
+      let(:relation) { ActiveFedora::Relation.new(ActiveFedora::Base) }
       it "should specify a q parameter" do
-        ActiveFedora::Base.should_receive(:find_one).with("changeme:30", true).and_return("Fake Object1")
-        ActiveFedora::Base.should_receive(:find_one).with("changeme:22", true).and_return("Fake Object2")
+        relation.should_receive(:find_one).with("changeme:30", true).and_return("Fake Object1")
+        relation.should_receive(:find_one).with("changeme:22", true).and_return("Fake Object2")
         mock_docs = [{"id" => "changeme:30"},{"id" => "changeme:22"}]
         mock_docs.should_receive(:has_next?).and_return(false)
         ActiveFedora::SolrService.instance.conn.should_receive(:paginate).with(1, 1000, 'select', :params=>{:q=>'*:*', :qt => 'standard', :sort => [@sort_query], :fl=> 'id', }).and_return('response'=>{'docs'=>mock_docs})
@@ -62,9 +65,14 @@ describe ActiveFedora::Base do
     end
 
     describe "with conditions" do
+      before do
+        ActiveFedora::Base.stub(:relation => relation)
+        relation.stub(clone: relation)
+      end
+      let(:relation) { ActiveFedora::Relation.new(SpecModel::Basic) }
       it "should filter by the provided fields" do
-        SpecModel::Basic.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
-        SpecModel::Basic.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
+        relation.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
+        relation.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
 
         mock_docs = [{"id" => "changeme:30"},{"id" => "changeme:22"}]
         mock_docs.should_receive(:has_next?).and_return(false)
@@ -92,8 +100,8 @@ describe ActiveFedora::Base do
       end
 
       it "should add options" do
-        SpecModel::Basic.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
-        SpecModel::Basic.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
+        relation.should_receive(:find_one).with("changeme:30", nil).and_return("Fake Object1")
+        relation.should_receive(:find_one).with("changeme:22", nil).and_return("Fake Object2")
 
         mock_docs = [{"id" => "changeme:30"},{"id" => "changeme:22"}]
         mock_docs.should_receive(:has_next?).and_return(false)
@@ -117,21 +125,23 @@ describe ActiveFedora::Base do
 
 
   describe '#find_each' do
+    before { ActiveFedora::Base.stub(:relation => relation) }
+    let(:relation) { ActiveFedora::Relation.new(SpecModel::Basic) }
     it "should query solr for all objects with :active_fedora_model_s of self.class" do
       mock_docs = [{"id" => "changeme:30"},{"id" => "changeme:22"}]
       mock_docs.should_receive(:has_next?).and_return(false)
       ActiveFedora::SolrService.instance.conn.should_receive(:paginate).with(1, 1000, 'select', :params=>{:q=>@model_query, :qt => 'standard', :sort => [@sort_query], :fl=> 'id', }).and_return('response'=>{'docs'=>mock_docs})
       
-      SpecModel::Basic.should_receive(:find_one).with("changeme:30", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:30'))
-      SpecModel::Basic.should_receive(:find_one).with("changeme:22", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:22'))
+      relation.should_receive(:find_one).with("changeme:30", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:30'))
+      relation.should_receive(:find_one).with("changeme:22", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:22'))
       yielded = double("yielded method")
       yielded.should_receive(:run).with { |obj| obj.class == SpecModel::Basic}.twice
       SpecModel::Basic.find_each(){|obj| yielded.run(obj) }
     end
     describe "with conditions" do
       it "should filter by the provided fields" do
-        SpecModel::Basic.should_receive(:find_one).with("changeme:30", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:30'))
-        SpecModel::Basic.should_receive(:find_one).with("changeme:22", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:22'))
+        relation.should_receive(:find_one).with("changeme:30", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:30'))
+        relation.should_receive(:find_one).with("changeme:22", nil).and_return(SpecModel::Basic.new(:pid=>'changeme:22'))
 
         mock_docs = [{"id" => "changeme:30"},{"id" => "changeme:22"}]
         mock_docs.should_receive(:has_next?).and_return(false)
